ExpandingList element made more self-contained
Added self variable to constructor and modified element
queries to base off self instead of document.
Element queries for lis and uls were run on the document
variable, and would modify ULs that were not children
of the customized ul element, and even modify ULs that were
did not specify is="expanding-list"

Merged creation of spans and the adding for click handler
and decoration to the spans.

Eliminated setting window.onload

Added a ul to index.html that is NOT and ExpandingList to show
that it is not affected by the code in the web component
definition

Added numerous comments explaining code purpose.